requirements.txt
opencv-contrib-python>=4.9.0.80
numpy>=1.26
PyYAML>=6.0.1
scipy>=1.11
rich>=13.7
